More code cleanup

+++ lib/src/shared/test/java/com/couchbase/lite/BaseTest.java
@@ -19,10 +19,8 @@ package com.couchbase.lite;
 
 import java.io.BufferedReader;
 import java.io.File;
-import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.util.Locale;
-import java.util.Map;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicReference;
@@ -30,8 +28,6 @@ import java.util.concurrent.atomic.AtomicReference;
 import org.json.JSONObject;
 import org.junit.After;
 import org.junit.Before;
-import org.junit.Rule;
-import org.junit.rules.ExpectedException;
 
 import com.couchbase.lite.internal.CouchbaseLiteInternal;
 import com.couchbase.lite.internal.ExecutionService;
@@ -42,7 +38,6 @@ import com.couchbase.lite.utils.Report;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
@@ -54,17 +49,13 @@ public abstract class BaseTest extends PlatformBaseTest {
     private static final int BUSY_RETRIES = 5;
 
     @FunctionalInterface
-    interface DocValidator extends Fn.ConsumerThrows<Document, CouchbaseLiteException> {}
-
-    private AtomicReference<AssertionError> testFailure;
+    public interface DocValidator extends Fn.ConsumerThrows<Document, CouchbaseLiteException> {}
 
     protected Database db;
     protected ExecutionService.CloseableExecutor executor;
 
+    private AtomicReference<AssertionError> testFailure;
     private File dbDir;
-
-    @Rule
-    public final ExpectedException thrown = ExpectedException.none();
 
     @Before
     public void setUp() throws CouchbaseLiteException {
@@ -74,92 +65,45 @@ public abstract class BaseTest extends PlatformBaseTest {
         //setupFileLogging(); // if needed
 
         executor = CouchbaseLiteInternal.getExecutionService().getSerialExecutor();
+
         testFailure = new AtomicReference<>();
 
         dbDir = new File(getDatabaseDirectoryPath());
         assertNotNull(dbDir);
 
         // if database exist, delete it
-        deleteDatabase(TEST_DB);
+        deleteDb(TEST_DB, dbDir);
 
         // clean dbDir
-        FileUtils.eraseFileOrDir(dbDir);
+        FileUtils.deleteContents(dbDir);
 
-        openDB();
+        db = new Database(TEST_DB);
     }
 
     @After
     public void tearDown() {
-        try {
-            closeDB();
-            deleteDatabase(TEST_DB);
-        }
+        try { eraseDatabase(db); }
         catch (CouchbaseLiteException e) {
             throw new RuntimeException("Failed closing database: " + TEST_DB, e);
         }
         finally {
-            if (dbDir != null) { FileUtils.eraseFileOrDir(dbDir); }
-
-            ExecutionService.CloseableExecutor exec = executor;
-            executor = null;
-            if (exec != null) { exec.stop(60, TimeUnit.SECONDS); }
-        }
-    }
-
-    protected Database openDB(String name) throws CouchbaseLiteException {
-        return new Database(name);
-    }
-
-    protected void openDB() throws CouchbaseLiteException {
-        assertNull(db);
-        db = openDB(TEST_DB);
-        assertNotNull(db);
-    }
-
-    protected void closeDB() throws CouchbaseLiteException {
-        closeDatabase(db);
-        db = null;
-    }
-
-    protected void closeDatabase(Database database) throws CouchbaseLiteException {
-        if (database == null) { return; }
-
-        for (int i = 0; i < BUSY_RETRIES; i++) {
             try {
-                database.close();
-                Report.log(LogLevel.VERBOSE, database.getName() + " was closed successfully.");
-                break;
+                if (dbDir != null) { FileUtils.eraseFileOrDir(dbDir); }
             }
-            catch (CouchbaseLiteException ex) {
-                if (ex.getCode() != CBLError.Code.BUSY) { throw ex; }
-                Report.log(LogLevel.VERBOSE, database.getName() + " cannot be closed as it is BUSY ...");
-                try { Thread.sleep(BUSY_WAIT_MS); }
-                catch (InterruptedException ignore) { }
+            finally {
+                stopExecutor();
             }
         }
     }
 
     protected void reopenDB() throws CouchbaseLiteException {
-        closeDB();
-        openDB();
+        closeDb(db);
+        db = new Database(TEST_DB);
     }
 
-    protected void cleanDB() throws CouchbaseLiteException {
-        if (db != null) {
-            db.delete();
-            db = null;
-        }
-        openDB();
-    }
-
-    protected void eraseDatabase(Database db) throws CouchbaseLiteException {
-        final String dbName = db.getName();
-        final File dbDir = db.getFilePath();
-        if (db.isOpen()) { closeDatabase(db); }
-        if (!deleteDatabase(dbName, dbDir.getParentFile())) {
-            Report.log(LogLevel.ERROR, "Failed to delete db: " + dbName + " @" + dbDir);
-        }
-        FileUtils.deleteContents(dbDir);
+    protected void recreateDB() throws CouchbaseLiteException {
+        if (db != null) { db.delete(); }
+        db = new Database(TEST_DB);
     }
 
     protected void deleteDatabase(Database db) throws CouchbaseLiteException {
@@ -171,74 +115,81 @@ public abstract class BaseTest extends PlatformBaseTest {
     }
 
     protected boolean deleteDatabase(String dbName, File dbDir) throws CouchbaseLiteException {
-        // database exist, delete it
-        if ((dbDir == null) || !dbDir.exists() || !Database.exists(dbName, dbDir)) { return true; }
+        return deleteDb(dbName, dbDir);
+    }
 
-        // If a test involves a replicator or a live query,
-        // it may take a while for the db to close
-        int i = 0;
-        while (true) {
-            try {
-                Database.delete(dbName, dbDir);
-                Report.log(LogLevel.VERBOSE, dbName + " was deleted successfully.");
-                return true;
-            }
-            catch (CouchbaseLiteException ex) {
-                if (ex.getCode() != CBLError.Code.BUSY) { throw ex; }
+    protected void eraseDatabase(Database db) throws CouchbaseLiteException {
+        if (db == null) { return; }
 
-                if (i++ >= BUSY_RETRIES) { return false; }
+        final String dbName = db.getName();
+        final File dbDir = db.getFilePath();
 
-                Report.log(LogLevel.WARNING, dbName + " cannot be deleted because it is BUSY ...");
-                try { Thread.sleep(BUSY_WAIT_MS); }
-                catch (InterruptedException ignore) { }
+        try {
+            if (!closeDb(db)) {
+                Report.log(LogLevel.ERROR, "Failed to close db: " + dbName + " @" + dbDir);
             }
         }
+        finally {
+            if (dbDir != null) {
+                try {
+                    if (!deleteDb(dbName, dbDir.getParentFile())) {
+                        Report.log(LogLevel.ERROR, "Failed to delete db: " + dbName + " @" + dbDir);
+                    }
+                }
+                finally {
+                    FileUtils.eraseFileOrDir(dbDir);
+                }
+            }
+        }
+    }
+
+    protected void loadJSONResource(String name) throws Exception {
+        try (BufferedReader in = new BufferedReader(new InputStreamReader(getAsset(name)))) {
+            int n = 1;
+            String line;
+            while ((line = in.readLine()) != null) {
+                if (line.trim().isEmpty()) { continue; }
+
+                MutableDocument doc = new MutableDocument(String.format(Locale.ENGLISH, "doc-%03d", n++));
+                doc.setData(JsonUtils.fromJson(new JSONObject(line)));
+
+                save(doc);
+            }
+        }
+    }
+
+    protected Document generateDocument(String docID) throws CouchbaseLiteException {
+        long n = db.getCount();
+
+        MutableDocument doc = new MutableDocument(docID);
+        doc.setValue("key", 1);
+        save(doc);
+        assertTrue((n + 1) == db.getCount());
+
+        Document savedDoc = db.getDocument(docID);
+        assertEquals(1, savedDoc.getSequence());
+
+        return savedDoc;
+    }
+
+    protected Document save(MutableDocument doc, DocValidator validator) throws CouchbaseLiteException {
+        validator.accept(doc);
+
+        Document savedDoc = save(doc);
+        validator.accept(doc);
+        validator.accept(savedDoc);
+
+        return savedDoc;
     }
 
     protected Document save(MutableDocument doc) throws CouchbaseLiteException {
         db.save(doc);
+
         Document savedDoc = db.getDocument(doc.getId());
         assertNotNull(savedDoc);
         assertEquals(doc.getId(), savedDoc.getId());
-        return savedDoc;
-    }
 
-    protected Document save(MutableDocument doc, DocValidator validator)
-        throws CouchbaseLiteException {
-        validator.accept(doc);
-        db.save(doc);
-        Document savedDoc = db.getDocument(doc.getId());
-        validator.accept(doc);
-        validator.accept(savedDoc);
         return savedDoc;
-    }
-
-    // helper method to save document
-    protected Document generateDocument(String docID) throws CouchbaseLiteException {
-        MutableDocument doc = new MutableDocument(docID);
-        doc.setValue("key", 1);
-        save(doc);
-        Document savedDoc = db.getDocument(docID);
-        assertTrue(db.getCount() > 0);
-        assertEquals(1, savedDoc.getSequence());
-        return savedDoc;
-    }
-
-    protected void loadJSONResource(String name) throws Exception {
-        InputStream is = getAsset(name);
-        BufferedReader reader = new BufferedReader(new InputStreamReader(is));
-        String line;
-        int n = 0;
-        while ((line = reader.readLine()) != null) {
-            if (line.trim().isEmpty()) { continue; }
-            n += 1;
-            JSONObject json = new JSONObject(line);
-            Map<String, Object> props = JsonUtils.fromJson(json);
-            String docId = String.format(Locale.ENGLISH, "doc-%03d", n);
-            MutableDocument doc = new MutableDocument(docId);
-            doc.setData(props);
-            save(doc);
-        }
     }
 
     protected void runSafely(Runnable test) {
@@ -274,5 +225,60 @@ public abstract class BaseTest extends PlatformBaseTest {
             assertEquals(domain, e.getDomain());
             assertEquals(code, e.getCode());
         }
+    }
+
+    private boolean closeDb(Database db) throws CouchbaseLiteException {
+        if ((db == null) || (!db.isOpen())) { return true; }
+
+        final String dbName = db.getName();
+
+        int i = 0;
+        while (true) {
+            try {
+                db.close();
+                Report.log(LogLevel.VERBOSE, dbName + " was deleted successfully.");
+                return true;
+            }
+            catch (CouchbaseLiteException ex) {
+                if (ex.getCode() != CBLError.Code.BUSY) { throw ex; }
+
+                if (i++ >= BUSY_RETRIES) { return false; }
+
+                Report.log(LogLevel.WARNING, dbName + " cannot be closed because it is busy...");
+                try { Thread.sleep(BUSY_WAIT_MS); }
+                catch (InterruptedException ignore) { return false; }
+            }
+        }
+    }
+
+    private boolean deleteDb(String dbName, File dbDir) throws CouchbaseLiteException {
+        // database exist, delete it
+        if ((dbDir == null) || !dbDir.exists() || !Database.exists(dbName, dbDir)) { return true; }
+
+        // If a test involves a replicator or a live query,
+        // it may take a while for the db to close
+        int i = 0;
+        while (true) {
+            try {
+                Database.delete(dbName, dbDir);
+                Report.log(LogLevel.VERBOSE, dbName + " was deleted successfully.");
+                return true;
+            }
+            catch (CouchbaseLiteException ex) {
+                if (ex.getCode() != CBLError.Code.BUSY) { throw ex; }
+
+                if (i++ >= BUSY_RETRIES) { return false; }
+
+                Report.log(LogLevel.WARNING, dbName + " cannot be deleted because it is busy...");
+                try { Thread.sleep(BUSY_WAIT_MS); }
+                catch (InterruptedException e) { return false; }
+            }
+        }
+    }
+
+    private void stopExecutor() {
+        ExecutionService.CloseableExecutor exec = executor;
+        executor = null;
+        if (exec != null) { exec.stop(60, TimeUnit.SECONDS); }
     }
 }
